+++ src/types/enums.ts
@@ -1,0 +1,33 @@
+export enum CountryCode {
+  IT = "IT",
+  US = "US",
+  DE = "DE",
+  FR = "FR",
+  GB = "GB",
+}
+
+export enum TaxIdType {
+  TAX_CODE = "Tax Code",
+  VAT_NUMBER = "VAT Number",
+  EIN = "EIN",
+  VAT_EU = "VAT EU",
+  OTHER = "Other",
+}
+
+export enum Day {
+  Mon = "Mon",
+  Tue = "Tue",
+  Wed = "Wed",
+  Thu = "Thu",
+  Fri = "Fri",
+  Sat = "Sat",
+  Sun = "Sun",
+}
+
+export enum ZoneType {
+  RURAL = "rural",
+  URBAN = "urban",
+  HOSPITALS = "hospitals",
+  MILITARY = "military",
+  RESTRICTED = "restricted",
+}